requirements.txt
cassandra-driver==2.5.0
cdeploy==1.1
paramiko==1.10
fabric==1.8.4
Flask==0.9
gunicorn==19.0.0
requests==2.6.0
google-api-python-client==1.4.0
blist==1.3.6
python-dateutil
